chore(core-terms-and-conditions): move StyledClickable to Chevron
